requirements.txt
ddgs
beautifulsoup4>=4.12.0
playwright>=1.44.0
schedule>=1.2.2
lxml>=5.2.0
python-dotenv>=1.0.0